chore: Remove dataElementService from web layer [DHIS2-1570]
