+++ dhis-2/dhis-web-api/src/test/java/org/hisp/dhis/webapi/controller/tracker/export/event/EventRequestParamsMapperTest.java
@@ -34,7 +34,6 @@ import static org.hisp.dhis.utils.Assertions.assertIsEmpty;
 import static org.hisp.dhis.utils.Assertions.assertStartsWith;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verifyNoInteractions;
@@ -42,14 +41,9 @@ import static org.mockito.Mockito.when;
 
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.hisp.dhis.common.AssignedUserSelectionMode;
-import org.hisp.dhis.common.QueryFilter;
-import org.hisp.dhis.common.QueryItem;
-import org.hisp.dhis.common.QueryOperator;
 import org.hisp.dhis.dataelement.DataElement;
 import org.hisp.dhis.dataelement.DataElementService;
 import org.hisp.dhis.feedback.BadRequestException;
@@ -74,7 +68,6 @@ import org.hisp.dhis.webapi.controller.event.webrequest.OrderCriteria;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.function.Executable;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -453,76 +446,5 @@ class EventRequestParamsMapperTest
             // contains
             () -> assertContains( "unsupportedProperty1", exception.getMessage() ),
             () -> assertContains( "unsupportedProperty2", exception.getMessage() ) );
-    }
-
-    @Test
-    void testFilter()
-        throws BadRequestException
-    {
-        RequestParams requestParams = new RequestParams();
-        requestParams.setFilter( Set.of( DE_1_UID + ":eq:2", DE_2_UID + ":like:foo" ) );
-
-        EventOperationParams params = mapper.map( requestParams );
-
-        List<QueryItem> items = params.getFilters();
-        assertNotNull( items );
-        // mapping to UIDs as the error message by just relying on QueryItem
-        // equals() is not helpful
-        assertContainsOnly( List.of( DE_1_UID,
-            DE_2_UID ), items.stream().map( i -> i.getItem().getUid() ).collect( Collectors.toList() ) );
-
-        // QueryItem equals() does not take the QueryFilter into account so
-        // assertContainsOnly alone does not ensure operators and filter value
-        // are correct
-        // the following block is needed because of that
-        // assertion is order independent as the order of QueryItems is not
-        // guaranteed
-        Map<String, QueryFilter> expectedFilters = Map.of(
-            DE_1_UID, new QueryFilter( QueryOperator.EQ, "2" ),
-            DE_2_UID, new QueryFilter( QueryOperator.LIKE, "foo" ) );
-        assertAll( items.stream().map( i -> (Executable) () -> {
-            String uid = i.getItem().getUid();
-            QueryFilter expected = expectedFilters.get( uid );
-            assertEquals( expected.getOperator().getValue() + " " + expected.getFilter(), i.getFiltersAsString(),
-                () -> String.format( "QueryFilter mismatch for DE with UID %s", uid ) );
-        } ).collect( Collectors.toList() ) );
-    }
-
-    @Test
-    void testFilterWhenDEHasMultipleFilters()
-        throws BadRequestException
-    {
-        RequestParams requestParams = new RequestParams();
-        requestParams.setFilter( Set.of( DE_1_UID + ":gt:10:lt:20" ) );
-
-        EventOperationParams params = mapper.map( requestParams );
-
-        List<QueryItem> items = params.getFilters();
-        assertNotNull( items );
-        // mapping to UIDs as the error message by just relying on QueryItem
-        // equals() is not helpful
-        assertContainsOnly( List.of( DE_1_UID ),
-            items.stream().map( i -> i.getItem().getUid() ).collect( Collectors.toList() ) );
-
-        // QueryItem equals() does not take the QueryFilter into account so
-        // assertContainsOnly alone does not ensure operators and filter value
-        // are correct
-        assertContainsOnly( Set.of(
-            new QueryFilter( QueryOperator.GT, "10" ),
-            new QueryFilter( QueryOperator.LT, "20" ) ), items.get( 0 ).getFilters() );
-    }
-
-    @Test
-    void shouldFailWithBadRequestExceptionWhenCriteriaDataElementDoesNotExist()
-    {
-        RequestParams requestParams = new RequestParams();
-        String filterName = "filter";
-        requestParams.setFilter( Set.of( filterName ) );
-        when( dataElementService.getDataElement( filterName ) ).thenReturn( null );
-
-        Exception exception = assertThrows( BadRequestException.class,
-            () -> mapper.map( requestParams ) );
-
-        assertEquals( "Data element does not exist: " + filterName, exception.getMessage() );
     }
 }
